Extract Voting Results for Honolulu City Council Measures

Extract Voting Results for Honolulu City Council Measures

Usage

get_votes(measures)

Arguments

measures

A tibble data frame of measures downloaded by `get_council_measures`

Value

A tibble data frame in long format with one row per council member per voting event.

Details

This function creates a long tibble data frame with one row per voting event per Honolulu City Council member who participated in the voting event. The output of this function excludes non-voting events such as the introduction of a measure or publishing public hearing notices.

Examples

measures_df <- get_council_measures(start_measure = 2796, end_measure = 2797)
votes_df <- get_votes(measures_df)